Add missing validation for checkout-complete with transactions
